Fix issues (#479)
* Checkin template for better way to work

* Update template

* Update scripts and gitignore

* Update support scripts

* Prepare init script

* Add stdio attachments

* Fix sdtio error by using duct

* Add waiton

* Update workspaces

* Update submodules

* Use workspace to test template

* Add tracing

* Update regex

+++ crates/init/src/file_utils.rs
@@ -26,11 +26,12 @@ pub async fn ensure_empty_folder<P: AsRef<Path>>(path: P) -> Result<()> {
 
 pub async fn delete_path<P: AsRef<Path>>(path: P) -> Result<()> {
     let path = path.as_ref();
-
-    if path.is_dir() {
-        fs::remove_dir_all(path).await?;
-    } else {
-        fs::remove_file(path).await?;
+    if path.exists() {
+        if path.is_dir() {
+            fs::remove_dir_all(path).await?;
+        } else {
+            fs::remove_file(path).await?;
+        }
     }
 
     Ok(())
@@ -41,6 +42,15 @@ pub async fn chmod_recursive<P: AsRef<Path>>(path: P, mode: &str) -> Result<()> 
         .arg("-R")
         .arg(mode)
         .arg(path.as_ref())
+        .status()
+        .await?;
+    Ok(())
+}
+
+pub async fn move_file<P: AsRef<Path>, Q: AsRef<Path>>(src: P, dst: Q) -> Result<()> {
+    Command::new("mv")
+        .arg(src.as_ref())
+        .arg(dst.as_ref())
         .status()
         .await?;
     Ok(())
